requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
h11==0.14.0
idna==3.7
lxml==5.2.2
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
python-docx==1.1.2
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
wsproto==1.2.0
